% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AcceleratorConfig}
\alias{AcceleratorConfig}
\title{Compute Engine API Objects 
Creates and runs virtual machines on Google Cloud Platform.}
\usage{
AcceleratorConfig(acceleratorCount = NULL, acceleratorType = NULL)
}
\arguments{
\item{acceleratorCount}{The number of the guest accelerator cards exposed to this instance}

\item{acceleratorType}{Full or partial URL of the accelerator type resource to expose to this instance}
}
\value{
AcceleratorConfig object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:37:45
filename: /Users/mark/dev/R/autoGoogleAPI/googlecomputebeta.auto/R/compute_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AcceleratorConfig Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A specification of the type and number of accelerator cards attached to the instance.
}
