% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AcceleratorType}
\alias{AcceleratorType}
\title{AcceleratorType Object}
\usage{
AcceleratorType(creationTimestamp = NULL, deprecated = NULL,
  description = NULL, id = NULL, maximumCardsPerInstance = NULL,
  name = NULL, selfLink = NULL, zone = NULL)
}
\arguments{
\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{deprecated}{[Output Only] The deprecation status associated with this accelerator type}

\item{description}{[Output Only] An optional textual description of the resource}

\item{id}{[Output Only] The unique identifier for the resource}

\item{maximumCardsPerInstance}{[Output Only] Maximum accelerator cards allowed per instance}

\item{name}{[Output Only] Name of the resource}

\item{selfLink}{[Output Only] Server-defined fully-qualified URL for this resource}

\item{zone}{[Output Only] The name of the zone where the accelerator type resides, such as us-central1-a}
}
\value{
AcceleratorType object
}
\description{
AcceleratorType Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Accelerator Type resource.
}
