% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AcceleratorTypeAggregatedList.items}
\alias{AcceleratorTypeAggregatedList.items}
\title{AcceleratorTypeAggregatedList.items Object}
\usage{
AcceleratorTypeAggregatedList.items()
}
\value{
AcceleratorTypeAggregatedList.items object
}
\description{
AcceleratorTypeAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] A map of scoped accelerator type lists.
}
\seealso{
Other AcceleratorTypeAggregatedList functions: \code{\link{AcceleratorTypeAggregatedList}}
}
