% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AccessConfig}
\alias{AccessConfig}
\title{AccessConfig Object}
\usage{
AccessConfig(name = NULL, natIP = NULL, type = NULL)
}
\arguments{
\item{name}{Name of this access configuration}

\item{natIP}{An external IP address associated with this instance}

\item{type}{The type of configuration}
}
\value{
AccessConfig object
}
\description{
AccessConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An access configuration attached to an instance's network interface. Only one access config per instance is supported.
}
\seealso{
Other AccessConfig functions: \code{\link{instances.addAccessConfig}}
}
