% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Address}
\alias{Address}
\title{Address Object}
\usage{
Address(address = NULL, creationTimestamp = NULL, description = NULL,
  id = NULL, name = NULL, region = NULL, selfLink = NULL,
  status = NULL, users = NULL)
}
\arguments{
\item{address}{The static external IP address represented by this resource}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{id}{[Output Only] The unique identifier for the resource}

\item{name}{Name of the resource}

\item{region}{[Output Only] URL of the region where the regional address resides}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{status}{[Output Only] The status of the address, which can be either IN_USE or RESERVED}

\item{users}{[Output Only] The URLs of the resources that are using this address}
}
\value{
Address object
}
\description{
Address Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A reserved address resource.
}
\seealso{
Other Address functions: \code{\link{addresses.insert}},
  \code{\link{globalAddresses.insert}}
}
