% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AutoscalersScopedList}
\alias{AutoscalersScopedList}
\title{AutoscalersScopedList Object}
\usage{
AutoscalersScopedList(AutoscalersScopedList.warning = NULL,
  AutoscalersScopedList.warning.data = NULL, autoscalers = NULL,
  warning = NULL)
}
\arguments{
\item{AutoscalersScopedList.warning}{The \link{AutoscalersScopedList.warning} object or list of objects}

\item{AutoscalersScopedList.warning.data}{The \link{AutoscalersScopedList.warning.data} object or list of objects}

\item{autoscalers}{[Output Only] List of autoscalers contained in this scope}

\item{warning}{[Output Only] Informational warning which replaces the list of autoscalers when the list is empty}
}
\value{
AutoscalersScopedList object
}
\description{
AutoscalersScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AutoscalersScopedList functions: \code{\link{AutoscalersScopedList.warning.data}},
  \code{\link{AutoscalersScopedList.warning}}
}

