% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AutoscalingPolicyCpuUtilization}
\alias{AutoscalingPolicyCpuUtilization}
\title{AutoscalingPolicyCpuUtilization Object}
\usage{
AutoscalingPolicyCpuUtilization(utilizationTarget = NULL)
}
\arguments{
\item{utilizationTarget}{The target CPU utilization that the autoscaler should maintain}
}
\value{
AutoscalingPolicyCpuUtilization object
}
\description{
AutoscalingPolicyCpuUtilization Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
CPU utilization policy.
}
