% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AutoscalingPolicyCustomMetricUtilization}
\alias{AutoscalingPolicyCustomMetricUtilization}
\title{AutoscalingPolicyCustomMetricUtilization Object}
\usage{
AutoscalingPolicyCustomMetricUtilization(metric = NULL,
  utilizationTarget = NULL, utilizationTargetType = NULL)
}
\arguments{
\item{metric}{The identifier of the Cloud Monitoring metric}

\item{utilizationTarget}{Target value of the metric which autoscaler should maintain}

\item{utilizationTargetType}{Defines how target utilization value is expressed for a Cloud Monitoring metric}
}
\value{
AutoscalingPolicyCustomMetricUtilization object
}
\description{
AutoscalingPolicyCustomMetricUtilization Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Custom utilization metric policy.
}

