% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{BackendBucketList}
\alias{BackendBucketList}
\title{BackendBucketList Object}
\usage{
BackendBucketList(id = NULL, items = NULL, nextPageToken = NULL,
  selfLink = NULL)
}
\arguments{
\item{id}{[Output Only] Unique identifier for the resource; defined by the server}

\item{items}{A list of BackendBucket resources}

\item{nextPageToken}{[Output Only] A token used to continue a truncated list request}

\item{selfLink}{[Output Only] Server-defined URL for this resource}
}
\value{
BackendBucketList object
}
\description{
BackendBucketList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains a list of BackendBucket resources.
}
