% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{BackendServiceAggregatedList}
\alias{BackendServiceAggregatedList}
\title{BackendServiceAggregatedList Object}
\usage{
BackendServiceAggregatedList(BackendServiceAggregatedList.items = NULL,
  id = NULL, items = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{BackendServiceAggregatedList.items}{The \link{BackendServiceAggregatedList.items} object or list of objects}

\item{id}{[Output Only] Unique identifier for the resource; defined by the server}

\item{items}{A map of scoped BackendService lists}

\item{nextPageToken}{[Output Only] A token used to continue a truncated list request}

\item{selfLink}{[Output Only] Server-defined URL for this resource}
}
\value{
BackendServiceAggregatedList object
}
\description{
BackendServiceAggregatedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains a list of BackendServicesScopedList.
}
\seealso{
Other BackendServiceAggregatedList functions: \code{\link{BackendServiceAggregatedList.items}}
}
