% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{BackendServiceCdnPolicy}
\alias{BackendServiceCdnPolicy}
\title{BackendServiceCdnPolicy Object}
\usage{
BackendServiceCdnPolicy(cacheKeyPolicy = NULL)
}
\arguments{
\item{cacheKeyPolicy}{The CacheKeyPolicy for this CdnPolicy}
}
\value{
BackendServiceCdnPolicy object
}
\description{
BackendServiceCdnPolicy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Message containing Cloud CDN configuration for a backend service.
}
