% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{BackendServiceIAP}
\alias{BackendServiceIAP}
\title{BackendServiceIAP Object}
\usage{
BackendServiceIAP(enabled = NULL, oauth2ClientId = NULL,
  oauth2ClientSecret = NULL, oauth2ClientSecretSha256 = NULL)
}
\arguments{
\item{enabled}{No description}

\item{oauth2ClientId}{No description}

\item{oauth2ClientSecret}{No description}

\item{oauth2ClientSecretSha256}{No description}
}
\value{
BackendServiceIAP object
}
\description{
BackendServiceIAP Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Identity-Aware Proxy (Cloud Gatekeeper)
}
