% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{CacheInvalidationRule}
\alias{CacheInvalidationRule}
\title{CacheInvalidationRule Object}
\usage{
CacheInvalidationRule(path = NULL)
}
\arguments{
\item{path}{No description}
}
\value{
CacheInvalidationRule object
}
\description{
CacheInvalidationRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other CacheInvalidationRule functions: \code{\link{urlMaps.invalidateCache}}
}

