% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{CustomerEncryptionKeyProtectedDisk}
\alias{CustomerEncryptionKeyProtectedDisk}
\title{CustomerEncryptionKeyProtectedDisk Object}
\usage{
CustomerEncryptionKeyProtectedDisk(diskEncryptionKey = NULL, source = NULL)
}
\arguments{
\item{diskEncryptionKey}{Decrypts data associated with the disk with a customer-supplied encryption key}

\item{source}{Specifies a valid partial or full URL to an existing Persistent Disk resource}
}
\value{
CustomerEncryptionKeyProtectedDisk object
}
\description{
CustomerEncryptionKeyProtectedDisk Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

