% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Disk.labels}
\alias{Disk.labels}
\title{Disk.labels Object}
\usage{
Disk.labels()
}
\value{
Disk.labels object
}
\description{
Disk.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Labels to apply to this disk. These can be later modified by the setLabels method. Each label key/value pair must comply with RFC1035. Label values may be empty.
}
\seealso{
Other Disk functions: \code{\link{Disk}},
  \code{\link{disks.insert}}
}

