% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Disk}
\alias{Disk}
\title{Disk Object}
\usage{
Disk(Disk.labels = NULL, creationTimestamp = NULL, description = NULL,
  diskEncryptionKey = NULL, id = NULL, labelFingerprint = NULL,
  labels = NULL, lastAttachTimestamp = NULL, lastDetachTimestamp = NULL,
  licenses = NULL, name = NULL, options = NULL, selfLink = NULL,
  sizeGb = NULL, sourceImage = NULL, sourceImageEncryptionKey = NULL,
  sourceImageId = NULL, sourceSnapshot = NULL,
  sourceSnapshotEncryptionKey = NULL, sourceSnapshotId = NULL,
  status = NULL, storageType = NULL, type = NULL, users = NULL,
  zone = NULL)
}
\arguments{
\item{Disk.labels}{The \link{Disk.labels} object or list of objects}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{diskEncryptionKey}{Encrypts the disk using a customer-supplied encryption key}

\item{id}{[Output Only] The unique identifier for the resource}

\item{labelFingerprint}{A fingerprint for the labels being applied to this disk, which is essentially a hash of the labels set used for optimistic locking}

\item{labels}{Labels to apply to this disk}

\item{lastAttachTimestamp}{[Output Only] Last attach timestamp in RFC3339 text format}

\item{lastDetachTimestamp}{[Output Only] Last detach timestamp in RFC3339 text format}

\item{licenses}{Any applicable publicly visible licenses}

\item{name}{Name of the resource}

\item{options}{Internal use only}

\item{selfLink}{[Output Only] Server-defined fully-qualified URL for this resource}

\item{sizeGb}{Size of the persistent disk, specified in GB}

\item{sourceImage}{The source image used to create this disk}

\item{sourceImageEncryptionKey}{The customer-supplied encryption key of the source image}

\item{sourceImageId}{[Output Only] The ID value of the image used to create this disk}

\item{sourceSnapshot}{The source snapshot used to create this disk}

\item{sourceSnapshotEncryptionKey}{The customer-supplied encryption key of the source snapshot}

\item{sourceSnapshotId}{[Output Only] The unique ID of the snapshot used to create this disk}

\item{status}{[Output Only] The status of disk creation}

\item{storageType}{[Deprecated] Storage type of the persistent disk}

\item{type}{URL of the disk type resource describing which disk type to use to create the disk}

\item{users}{[Output Only] Links to the users of the disk (attached instances) in form: project/zones/zone/instances/instance}

\item{zone}{[Output Only] URL of the zone where the disk resides}
}
\value{
Disk object
}
\description{
Disk Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Disk resource.
}
\seealso{
Other Disk functions: \code{\link{Disk.labels}},
  \code{\link{disks.insert}}
}

