% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{DiskTypesScopedList}
\alias{DiskTypesScopedList}
\title{DiskTypesScopedList Object}
\usage{
DiskTypesScopedList(DiskTypesScopedList.warning = NULL,
  DiskTypesScopedList.warning.data = NULL, diskTypes = NULL,
  warning = NULL)
}
\arguments{
\item{DiskTypesScopedList.warning}{The \link{DiskTypesScopedList.warning} object or list of objects}

\item{DiskTypesScopedList.warning.data}{The \link{DiskTypesScopedList.warning.data} object or list of objects}

\item{diskTypes}{[Output Only] List of disk types contained in this scope}

\item{warning}{[Output Only] Informational warning which replaces the list of disk types when the list is empty}
}
\value{
DiskTypesScopedList object
}
\description{
DiskTypesScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other DiskTypesScopedList functions: \code{\link{DiskTypesScopedList.warning.data}},
  \code{\link{DiskTypesScopedList.warning}}
}

