% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{DiskTypesScopedList.warning}
\alias{DiskTypesScopedList.warning}
\title{DiskTypesScopedList.warning Object}
\usage{
DiskTypesScopedList.warning(DiskTypesScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{DiskTypesScopedList.warning.data}{The \link{DiskTypesScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
DiskTypesScopedList.warning object
}
\description{
DiskTypesScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] Informational warning which replaces the list of disk types when the list is empty.
}
\seealso{
Other DiskTypesScopedList functions: \code{\link{DiskTypesScopedList.warning.data}},
  \code{\link{DiskTypesScopedList}}
}

