% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ForwardingRule}
\alias{ForwardingRule}
\title{ForwardingRule Object}
\usage{
ForwardingRule(IPAddress = NULL, IPProtocol = NULL,
  creationTimestamp = NULL, description = NULL, id = NULL, name = NULL,
  portRange = NULL, region = NULL, selfLink = NULL, target = NULL)
}
\arguments{
\item{IPAddress}{The IP address that this forwarding rule is serving on behalf of}

\item{IPProtocol}{The IP protocol to which this rule applies}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{id}{[Output Only] The unique identifier for the resource}

\item{name}{Name of the resource; provided by the client when the resource is created}

\item{portRange}{Applicable only when IPProtocol is TCP, UDP, or SCTP, only packets addressed to ports in the specified range will be forwarded to target}

\item{region}{[Output Only] URL of the region where the regional forwarding rule resides}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{target}{The URL of the target resource to receive the matched traffic}
}
\value{
ForwardingRule object
}
\description{
ForwardingRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A ForwardingRule resource. A ForwardingRule resource specifies which pool of target virtual machines to forward a packet to if it matches the given [IPAddress, IPProtocol, portRange] tuple.
}
\seealso{
Other ForwardingRule functions: \code{\link{forwardingRules.insert}},
  \code{\link{globalForwardingRules.insert}}
}

