% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ForwardingRuleAggregatedList.items}
\alias{ForwardingRuleAggregatedList.items}
\title{ForwardingRuleAggregatedList.items Object}
\usage{
ForwardingRuleAggregatedList.items()
}
\value{
ForwardingRuleAggregatedList.items object
}
\description{
ForwardingRuleAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of scoped forwarding rule lists.
}
\seealso{
Other ForwardingRuleAggregatedList functions: \code{\link{ForwardingRuleAggregatedList}}
}

