% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ForwardingRulesScopedList}
\alias{ForwardingRulesScopedList}
\title{ForwardingRulesScopedList Object}
\usage{
ForwardingRulesScopedList(ForwardingRulesScopedList.warning = NULL,
  ForwardingRulesScopedList.warning.data = NULL, forwardingRules = NULL,
  warning = NULL)
}
\arguments{
\item{ForwardingRulesScopedList.warning}{The \link{ForwardingRulesScopedList.warning} object or list of objects}

\item{ForwardingRulesScopedList.warning.data}{The \link{ForwardingRulesScopedList.warning.data} object or list of objects}

\item{forwardingRules}{List of forwarding rules contained in this scope}

\item{warning}{Informational warning which replaces the list of forwarding rules when the list is empty}
}
\value{
ForwardingRulesScopedList object
}
\description{
ForwardingRulesScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ForwardingRulesScopedList functions: \code{\link{ForwardingRulesScopedList.warning.data}},
  \code{\link{ForwardingRulesScopedList.warning}}
}
