% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ForwardingRulesScopedList.warning}
\alias{ForwardingRulesScopedList.warning}
\title{ForwardingRulesScopedList.warning Object}
\usage{

  ForwardingRulesScopedList.warning(ForwardingRulesScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{ForwardingRulesScopedList.warning.data}{The \link{ForwardingRulesScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
ForwardingRulesScopedList.warning object
}
\description{
ForwardingRulesScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Informational warning which replaces the list of forwarding rules when the list is empty.
}
\seealso{
Other ForwardingRulesScopedList functions: \code{\link{ForwardingRulesScopedList.warning.data}},
  \code{\link{ForwardingRulesScopedList}}
}

