% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{HealthCheck}
\alias{HealthCheck}
\title{HealthCheck Object}
\usage{
HealthCheck(checkIntervalSec = NULL, creationTimestamp = NULL,
  description = NULL, healthyThreshold = NULL, httpHealthCheck = NULL,
  httpsHealthCheck = NULL, id = NULL, name = NULL, selfLink = NULL,
  sslHealthCheck = NULL, tcpHealthCheck = NULL, timeoutSec = NULL,
  type = NULL, udpHealthCheck = NULL, unhealthyThreshold = NULL)
}
\arguments{
\item{checkIntervalSec}{How often (in seconds) to send a health check}

\item{creationTimestamp}{[Output Only] Creation timestamp in 3339 text format}

\item{description}{An optional description of this resource}

\item{healthyThreshold}{A so-far unhealthy instance will be marked healthy after this many consecutive successes}

\item{httpHealthCheck}{No description}

\item{httpsHealthCheck}{No description}

\item{id}{[Output Only] The unique identifier for the resource}

\item{name}{Name of the resource}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{sslHealthCheck}{No description}

\item{tcpHealthCheck}{No description}

\item{timeoutSec}{How long (in seconds) to wait before claiming failure}

\item{type}{Specifies the type of the healthCheck, either TCP, SSL, HTTP or HTTPS}

\item{udpHealthCheck}{No description}

\item{unhealthyThreshold}{A so-far healthy instance will be marked unhealthy after this many consecutive failures}
}
\value{
HealthCheck object
}
\description{
HealthCheck Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An HealthCheck resource. This resource defines a template for how individual virtual machines should be checked for health, via one of the supported protocols.
}
\seealso{
Other HealthCheck functions: \code{\link{healthChecks.insert}},
  \code{\link{healthChecks.patch}},
  \code{\link{healthChecks.update}}
}
