% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{httpsHealthChecks.update}
\alias{httpsHealthChecks.update}
\title{Updates a HttpsHealthCheck resource in the specified project using the data included in the request.}
\usage{
httpsHealthChecks.update(HttpsHealthCheck, project, httpsHealthCheck)
}
\arguments{
\item{HttpsHealthCheck}{The \link{HttpsHealthCheck} object to pass to this method}

\item{project}{Project ID for this request}

\item{httpsHealthCheck}{Name of the HttpsHealthCheck resource to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other HttpsHealthCheck functions: \code{\link{HttpsHealthCheck}},
  \code{\link{httpsHealthChecks.insert}},
  \code{\link{httpsHealthChecks.patch}}
}
