% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Instance.labels}
\alias{Instance.labels}
\title{Instance.labels Object}
\usage{
Instance.labels()
}
\value{
Instance.labels object
}
\description{
Instance.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Labels to apply to this instance. These can be later modified by the setLabels method. Each label key/value pair must comply with RFC1035. Label values may be empty.
}
\seealso{
Other Instance functions: \code{\link{Instance}},
  \code{\link{instances.insert}}
}

