% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroup}
\alias{InstanceGroup}
\title{InstanceGroup Object}
\usage{
InstanceGroup(creationTimestamp = NULL, description = NULL,
  fingerprint = NULL, id = NULL, name = NULL, namedPorts = NULL,
  network = NULL, region = NULL, selfLink = NULL, size = NULL,
  subnetwork = NULL, zone = NULL)
}
\arguments{
\item{creationTimestamp}{[Output Only] The creation timestamp for this instance group in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{fingerprint}{[Output Only] The fingerprint of the named ports}

\item{id}{[Output Only] A unique identifier for this resource type}

\item{name}{The name of the instance group}

\item{namedPorts}{Assigns a name to a port number}

\item{network}{The URL of the network to which all instances in the instance group belong}

\item{region}{The URL of the region where the instance group is located (for regional resources)}

\item{selfLink}{[Output Only] The URL for this instance group}

\item{size}{[Output Only] The total number of instances in the instance group}

\item{subnetwork}{The URL of the subnetwork to which all instances in the instance group belong}

\item{zone}{[Output Only] The URL of the zone where the instance group is located (for zonal resources)}
}
\value{
InstanceGroup object
}
\description{
InstanceGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstanceGroup functions: \code{\link{instanceGroups.insert}}
}

