% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupAggregatedList}
\alias{InstanceGroupAggregatedList}
\title{InstanceGroupAggregatedList Object}
\usage{
InstanceGroupAggregatedList(InstanceGroupAggregatedList.items = NULL,
  id = NULL, items = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{InstanceGroupAggregatedList.items}{The \link{InstanceGroupAggregatedList.items} object or list of objects}

\item{id}{[Output Only] A unique identifier for this aggregated list of instance groups}

\item{items}{A map of scoped instance group lists}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] The URL for this resource type}
}
\value{
InstanceGroupAggregatedList object
}
\description{
InstanceGroupAggregatedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstanceGroupAggregatedList functions: \code{\link{InstanceGroupAggregatedList.items}}
}
