% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupManagerAggregatedList.items}
\alias{InstanceGroupManagerAggregatedList.items}
\title{InstanceGroupManagerAggregatedList.items Object}
\usage{
InstanceGroupManagerAggregatedList.items()
}
\value{
InstanceGroupManagerAggregatedList.items object
}
\description{
InstanceGroupManagerAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] A map of filtered managed instance group lists.
}
\seealso{
Other InstanceGroupManagerAggregatedList functions: \code{\link{InstanceGroupManagerAggregatedList}}
}

