% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupManagerAggregatedList}
\alias{InstanceGroupManagerAggregatedList}
\title{InstanceGroupManagerAggregatedList Object}
\usage{

  InstanceGroupManagerAggregatedList(InstanceGroupManagerAggregatedList.items = NULL,
  id = NULL, items = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{InstanceGroupManagerAggregatedList.items}{The \link{InstanceGroupManagerAggregatedList.items} object or list of objects}

\item{id}{[Output Only] A unique identifier for this aggregated list of managed instance groups}

\item{items}{[Output Only] A map of filtered managed instance group lists}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] The URL for this resource type}
}
\value{
InstanceGroupManagerAggregatedList object
}
\description{
InstanceGroupManagerAggregatedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstanceGroupManagerAggregatedList functions: \code{\link{InstanceGroupManagerAggregatedList.items}}
}

