% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupManagerAutoHealingPolicy}
\alias{InstanceGroupManagerAutoHealingPolicy}
\title{InstanceGroupManagerAutoHealingPolicy Object}
\usage{
InstanceGroupManagerAutoHealingPolicy(healthCheck = NULL,
  initialDelaySec = NULL)
}
\arguments{
\item{healthCheck}{The URL for the health check that signals autohealing}

\item{initialDelaySec}{The number of seconds that the managed instance group waits before it applies autohealing policies to new instances or recently recreated instances}
}
\value{
InstanceGroupManagerAutoHealingPolicy object
}
\description{
InstanceGroupManagerAutoHealingPolicy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
