% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupManagersSetAutoHealingRequest}
\alias{InstanceGroupManagersSetAutoHealingRequest}
\title{InstanceGroupManagersSetAutoHealingRequest Object}
\usage{
InstanceGroupManagersSetAutoHealingRequest(autoHealingPolicies = NULL)
}
\arguments{
\item{autoHealingPolicies}{No description}
}
\value{
InstanceGroupManagersSetAutoHealingRequest object
}
\description{
InstanceGroupManagersSetAutoHealingRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstanceGroupManagersSetAutoHealingRequest functions: \code{\link{instanceGroupManagers.setAutoHealingPolicies}}
}
