% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupsSetNamedPortsRequest}
\alias{InstanceGroupsSetNamedPortsRequest}
\title{InstanceGroupsSetNamedPortsRequest Object}
\usage{
InstanceGroupsSetNamedPortsRequest(fingerprint = NULL, namedPorts = NULL)
}
\arguments{
\item{fingerprint}{The fingerprint of the named ports information for this instance group}

\item{namedPorts}{The list of named ports to set for this instance group}
}
\value{
InstanceGroupsSetNamedPortsRequest object
}
\description{
InstanceGroupsSetNamedPortsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstanceGroupsSetNamedPortsRequest functions: \code{\link{instanceGroups.setNamedPorts}}
}
