% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceMoveRequest}
\alias{InstanceMoveRequest}
\title{InstanceMoveRequest Object}
\usage{
InstanceMoveRequest(destinationZone = NULL, targetInstance = NULL)
}
\arguments{
\item{destinationZone}{The URL of the destination zone to move the instance}

\item{targetInstance}{The URL of the target instance to move}
}
\value{
InstanceMoveRequest object
}
\description{
InstanceMoveRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstanceMoveRequest functions: \code{\link{projects.moveInstance}}
}

