% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceProperties}
\alias{InstanceProperties}
\title{InstanceProperties Object}
\usage{
InstanceProperties(InstanceProperties.labels = NULL, canIpForward = NULL,
  description = NULL, disks = NULL, labels = NULL, machineType = NULL,
  metadata = NULL, networkInterfaces = NULL, scheduling = NULL,
  serviceAccounts = NULL, tags = NULL)
}
\arguments{
\item{InstanceProperties.labels}{The \link{InstanceProperties.labels} object or list of objects}

\item{canIpForward}{Enables instances created based on this template to send packets with source IP addresses other than their own and receive packets with destination IP addresses other than their own}

\item{description}{An optional text description for the instances that are created from this instance template}

\item{disks}{An array of disks that are associated with the instances that are created from this template}

\item{labels}{Labels to apply to instances that are created from this template}

\item{machineType}{The machine type to use for instances that are created from this template}

\item{metadata}{The metadata key/value pairs to assign to instances that are created from this template}

\item{networkInterfaces}{An array of network access configurations for this interface}

\item{scheduling}{Specifies the scheduling options for the instances that are created from this template}

\item{serviceAccounts}{A list of service accounts with specified scopes}

\item{tags}{A list of tags to apply to the instances that are created from this template}
}
\value{
InstanceProperties object
}
\description{
InstanceProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}
\seealso{
Other InstanceProperties functions: \code{\link{InstanceProperties.labels}}
}

