% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceReference}
\alias{InstanceReference}
\title{InstanceReference Object}
\usage{
InstanceReference(instance = NULL)
}
\arguments{
\item{instance}{The URL for a specific instance}
}
\value{
InstanceReference object
}
\description{
InstanceReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstanceReference functions: \code{\link{targetPools.getHealth}}
}

