% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{instances.insert}
\alias{instances.insert}
\title{Creates an instance resource in the specified project using the data included in the request.}
\usage{
instances.insert(Instance, project, zone)
}
\arguments{
\item{Instance}{The \link{Instance} object to pass to this method}

\item{project}{Project ID for this request}

\item{zone}{The name of the zone for this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other Instance functions: \code{\link{Instance.labels}},
  \code{\link{Instance}}
}

