% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{instances.setServiceAccount}
\alias{instances.setServiceAccount}
\title{Sets the service account on the instance. For more information, read Changing the service account and access scopes for an instance.}
\usage{
instances.setServiceAccount(InstancesSetServiceAccountRequest, project, zone,
  instance)
}
\arguments{
\item{InstancesSetServiceAccountRequest}{The \link{InstancesSetServiceAccountRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{zone}{The name of the zone for this request}

\item{instance}{Name of the instance resource to start}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other InstancesSetServiceAccountRequest functions: \code{\link{InstancesSetServiceAccountRequest}}
}
