% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstancesScopedList.warning}
\alias{InstancesScopedList.warning}
\title{InstancesScopedList.warning Object}
\usage{
InstancesScopedList.warning(InstancesScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{InstancesScopedList.warning.data}{The \link{InstancesScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
InstancesScopedList.warning object
}
\description{
InstancesScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] Informational warning which replaces the list of instances when the list is empty.
}
\seealso{
Other InstancesScopedList functions: \code{\link{InstancesScopedList.warning.data}},
  \code{\link{InstancesScopedList}}
}
