% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstancesSetLabelsRequest.labels}
\alias{InstancesSetLabelsRequest.labels}
\title{InstancesSetLabelsRequest.labels Object}
\usage{
InstancesSetLabelsRequest.labels()
}
\value{
InstancesSetLabelsRequest.labels object
}
\description{
InstancesSetLabelsRequest.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstancesSetLabelsRequest functions: \code{\link{InstancesSetLabelsRequest}},
  \code{\link{instances.setLabels}}
}
