% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{instanceTemplates.insert}
\alias{instanceTemplates.insert}
\title{Creates an instance template in the specified project using the data that is included in the request. If you are creating a new template to update an existing instance group, your new instance template must use the same network or, if applicable, the same subnetwork as the original template.}
\usage{
instanceTemplates.insert(InstanceTemplate, project)
}
\arguments{
\item{InstanceTemplate}{The \link{InstanceTemplate} object to pass to this method}

\item{project}{Project ID for this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other InstanceTemplate functions: \code{\link{InstanceTemplate}}
}
