% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Metadata.items}
\alias{Metadata.items}
\title{Metadata.items Object}
\usage{
Metadata.items()
}
\value{
Metadata.items object
}
\description{
Metadata.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Metadata functions: \code{\link{Metadata}},
  \code{\link{instances.setMetadata}},
  \code{\link{projects.setCommonInstanceMetadata}}
}

