% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{NetworkInterface}
\alias{NetworkInterface}
\title{NetworkInterface Object}
\usage{
NetworkInterface(accessConfigs = NULL, name = NULL, network = NULL,
  networkIP = NULL, subnetwork = NULL)
}
\arguments{
\item{accessConfigs}{An array of configurations for this interface}

\item{name}{[Output Only] The name of the network interface, generated by the server}

\item{network}{URL of the network resource for this instance}

\item{networkIP}{An IPv4 internal network address to assign to the instance for this network interface}

\item{subnetwork}{The URL of the Subnetwork resource for this instance}
}
\value{
NetworkInterface object
}
\description{
NetworkInterface Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A network interface resource attached to an instance.
}
