% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Operation.error}
\alias{Operation.error}
\title{Operation.error Object}
\usage{
Operation.error(Operation.error.errors = NULL, errors = NULL)
}
\arguments{
\item{Operation.error.errors}{The \link{Operation.error.errors} object or list of objects}

\item{errors}{[Output Only] The array of errors encountered while processing this operation}
}
\value{
Operation.error object
}
\description{
Operation.error Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] If errors are generated during processing of the operation, this field will be populated.
}
\seealso{
Other Operation functions: \code{\link{Operation.error.errors}},
  \code{\link{Operation.warnings.data}},
  \code{\link{Operation.warnings}}, \code{\link{Operation}}
}

