% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{OperationsScopedList}
\alias{OperationsScopedList}
\title{OperationsScopedList Object}
\usage{
OperationsScopedList(OperationsScopedList.warning = NULL,
  OperationsScopedList.warning.data = NULL, operations = NULL,
  warning = NULL)
}
\arguments{
\item{OperationsScopedList.warning}{The \link{OperationsScopedList.warning} object or list of objects}

\item{OperationsScopedList.warning.data}{The \link{OperationsScopedList.warning.data} object or list of objects}

\item{operations}{[Output Only] List of operations contained in this scope}

\item{warning}{[Output Only] Informational warning which replaces the list of operations when the list is empty}
}
\value{
OperationsScopedList object
}
\description{
OperationsScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other OperationsScopedList functions: \code{\link{OperationsScopedList.warning.data}},
  \code{\link{OperationsScopedList.warning}}
}

