% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ProjectsListXpnHostsRequest}
\alias{ProjectsListXpnHostsRequest}
\title{ProjectsListXpnHostsRequest Object}
\usage{
ProjectsListXpnHostsRequest(organization = NULL)
}
\arguments{
\item{organization}{Optional organization ID managed by Cloud Resource Manager, for which to list XPN host projects}
}
\value{
ProjectsListXpnHostsRequest object
}
\description{
ProjectsListXpnHostsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ProjectsListXpnHostsRequest functions: \code{\link{projects.listXpnHosts}}
}
