% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RegionAutoscalerList}
\alias{RegionAutoscalerList}
\title{RegionAutoscalerList Object}
\usage{
RegionAutoscalerList(id = NULL, items = NULL, nextPageToken = NULL,
  selfLink = NULL)
}
\arguments{
\item{id}{[Output Only] The unique identifier for the resource}

\item{items}{A list of autoscalers}

\item{nextPageToken}{[Output Only] A token used to continue a truncated list request}

\item{selfLink}{[Output Only] Server-defined URL for this resource}
}
\value{
RegionAutoscalerList object
}
\description{
RegionAutoscalerList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains a list of autoscalers.
}

