% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RegionInstanceGroupManagerList}
\alias{RegionInstanceGroupManagerList}
\title{RegionInstanceGroupManagerList Object}
\usage{
RegionInstanceGroupManagerList(id = NULL, items = NULL,
  nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{id}{[Output Only] The unique identifier for the resource}

\item{items}{A list of managed instance groups}

\item{nextPageToken}{[Output only] A token used to continue a truncated list request}

\item{selfLink}{[Output only] The URL for this resource type}
}
\value{
RegionInstanceGroupManagerList object
}
\description{
RegionInstanceGroupManagerList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains a list of managed instance groups.
}
