% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{regionInstanceGroupManagers.deleteInstances}
\alias{regionInstanceGroupManagers.deleteInstances}
\title{Schedules a group action to delete the specified instances in the managed instance group. The instances are also removed from any target pools of which they were a member. This method reduces the targetSize of the managed instance group by the number of instances that you delete. This operation is marked as DONE when the action is scheduled even if the instances are still being deleted. You must separately verify the status of the deleting action with the listmanagedinstances method.}
\usage{

  regionInstanceGroupManagers.deleteInstances(RegionInstanceGroupManagersDeleteInstancesRequest,
  project, region, instanceGroupManager)
}
\arguments{
\item{RegionInstanceGroupManagersDeleteInstancesRequest}{The \link{RegionInstanceGroupManagersDeleteInstancesRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{region}{Name of the region scoping this request}

\item{instanceGroupManager}{Name of the managed instance group}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other RegionInstanceGroupManagersDeleteInstancesRequest functions: \code{\link{RegionInstanceGroupManagersDeleteInstancesRequest}}
}

