% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{regionInstanceGroupManagers.insert}
\alias{regionInstanceGroupManagers.insert}
\title{Creates a managed instance group using the information that you specify in the request. After the group is created, it schedules an action to create instances in the group using the specified instance template. This operation is marked as DONE when the group is created even if the instances in the group have not yet been created. You must separately verify the status of the individual instances with the listmanagedinstances method.}
\usage{
regionInstanceGroupManagers.insert(InstanceGroupManager, project, region)
}
\arguments{
\item{InstanceGroupManager}{The \link{InstanceGroupManager} object to pass to this method}

\item{project}{Project ID for this request}

\item{region}{Name of the region scoping this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other InstanceGroupManager functions: \code{\link{InstanceGroupManager}},
  \code{\link{instanceGroupManagers.insert}},
  \code{\link{instanceGroupManagers.patch}},
  \code{\link{instanceGroupManagers.update}},
  \code{\link{regionInstanceGroupManagers.patch}},
  \code{\link{regionInstanceGroupManagers.update}}
}
