% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RegionInstanceGroupManagersAbandonInstancesRequest}
\alias{RegionInstanceGroupManagersAbandonInstancesRequest}
\title{RegionInstanceGroupManagersAbandonInstancesRequest Object}
\usage{
RegionInstanceGroupManagersAbandonInstancesRequest(instances = NULL)
}
\arguments{
\item{instances}{The names of one or more instances to abandon}
}
\value{
RegionInstanceGroupManagersAbandonInstancesRequest object
}
\description{
RegionInstanceGroupManagersAbandonInstancesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other RegionInstanceGroupManagersAbandonInstancesRequest functions: \code{\link{regionInstanceGroupManagers.abandonInstances}}
}

