% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RegionInstanceGroupManagersSetAutoHealingRequest}
\alias{RegionInstanceGroupManagersSetAutoHealingRequest}
\title{RegionInstanceGroupManagersSetAutoHealingRequest Object}
\usage{
RegionInstanceGroupManagersSetAutoHealingRequest(autoHealingPolicies = NULL)
}
\arguments{
\item{autoHealingPolicies}{No description}
}
\value{
RegionInstanceGroupManagersSetAutoHealingRequest object
}
\description{
RegionInstanceGroupManagersSetAutoHealingRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other RegionInstanceGroupManagersSetAutoHealingRequest functions: \code{\link{regionInstanceGroupManagers.setAutoHealingPolicies}}
}
