% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{regionInstanceGroups.listInstances}
\alias{regionInstanceGroups.listInstances}
\title{Lists the instances in the specified instance group and displays information about the named ports. Depending on the specified options, this method can list all instances or only the instances that are running.}
\usage{
regionInstanceGroups.listInstances(RegionInstanceGroupsListInstancesRequest,
  project, region, instanceGroup, filter = NULL, maxResults = NULL,
  orderBy = NULL, pageToken = NULL)
}
\arguments{
\item{RegionInstanceGroupsListInstancesRequest}{The \link{RegionInstanceGroupsListInstancesRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{region}{Name of the region scoping this request}

\item{instanceGroup}{Name of the regional instance group for which we want to list the instances}

\item{filter}{Sets a filter expression for filtering listed resources, in the form filter={expression}}

\item{maxResults}{The maximum number of results per page that should be returned}

\item{orderBy}{Sorts list results by a certain order}

\item{pageToken}{Specifies a page token to use}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute, https://www.googleapis.com/auth/compute.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other RegionInstanceGroupsListInstancesRequest functions: \code{\link{RegionInstanceGroupsListInstancesRequest}}
}

