% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Router}
\alias{Router}
\title{Router Object}
\usage{
Router(bgp = NULL, bgpPeers = NULL, creationTimestamp = NULL,
  description = NULL, id = NULL, interfaces = NULL, name = NULL,
  network = NULL, region = NULL, selfLink = NULL)
}
\arguments{
\item{bgp}{BGP information specific to this router}

\item{bgpPeers}{BGP information that needs to be configured into the routing stack to establish the BGP peering}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{id}{[Output Only] The unique identifier for the resource}

\item{interfaces}{Router interfaces}

\item{name}{Name of the resource}

\item{network}{URI of the network to which this router belongs}

\item{region}{[Output Only] URI of the region where the router resides}

\item{selfLink}{[Output Only] Server-defined URL for the resource}
}
\value{
Router object
}
\description{
Router Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Router resource.
}
\seealso{
Other Router functions: \code{\link{routers.insert}},
  \code{\link{routers.patch}},
  \code{\link{routers.preview}},
  \code{\link{routers.update}}
}
