% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RouterAggregatedList.items}
\alias{RouterAggregatedList.items}
\title{RouterAggregatedList.items Object}
\usage{
RouterAggregatedList.items()
}
\value{
RouterAggregatedList.items object
}
\description{
RouterAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of scoped router lists.
}
\seealso{
Other RouterAggregatedList functions: \code{\link{RouterAggregatedList}}
}
