% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RouterInterface}
\alias{RouterInterface}
\title{RouterInterface Object}
\usage{
RouterInterface(ipRange = NULL, linkedVpnTunnel = NULL, name = NULL)
}
\arguments{
\item{ipRange}{IP address and range of the interface}

\item{linkedVpnTunnel}{URI of linked VPN tunnel}

\item{name}{Name of this interface entry}
}
\value{
RouterInterface object
}
\description{
RouterInterface Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
