% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RoutersScopedList.warning.data}
\alias{RoutersScopedList.warning.data}
\title{RoutersScopedList.warning.data Object}
\usage{
RoutersScopedList.warning.data()
}
\value{
RoutersScopedList.warning.data object
}
\description{
RoutersScopedList.warning.data Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other RoutersScopedList functions: \code{\link{RoutersScopedList.warning}},
  \code{\link{RoutersScopedList}}
}

