% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RouterStatus}
\alias{RouterStatus}
\title{RouterStatus Object}
\usage{
RouterStatus(bestRoutes = NULL, bgpPeerStatus = NULL, network = NULL)
}
\arguments{
\item{bestRoutes}{Best routes for this router's network}

\item{bgpPeerStatus}{No description}

\item{network}{URI of the network to which this router belongs}
}
\value{
RouterStatus object
}
\description{
RouterStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

