% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Scheduling}
\alias{Scheduling}
\title{Scheduling Object}
\usage{
Scheduling(automaticRestart = NULL, onHostMaintenance = NULL,
  preemptible = NULL)
}
\arguments{
\item{automaticRestart}{Specifies whether the instance should be automatically restarted if it is terminated by Compute Engine (not terminated by a user)}

\item{onHostMaintenance}{Defines the maintenance behavior for this instance}

\item{preemptible}{Whether the instance is preemptible}
}
\value{
Scheduling object
}
\description{
Scheduling Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Sets the scheduling options for an Instance.
}
\seealso{
Other Scheduling functions: \code{\link{instances.setScheduling}}
}
