% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Snapshot}
\alias{Snapshot}
\title{Snapshot Object}
\usage{
Snapshot(Snapshot.labels = NULL, creationTimestamp = NULL,
  description = NULL, diskSizeGb = NULL, id = NULL,
  labelFingerprint = NULL, labels = NULL, licenses = NULL, name = NULL,
  selfLink = NULL, snapshotEncryptionKey = NULL, sourceDisk = NULL,
  sourceDiskEncryptionKey = NULL, sourceDiskId = NULL, status = NULL,
  storageBytes = NULL, storageBytesStatus = NULL)
}
\arguments{
\item{Snapshot.labels}{The \link{Snapshot.labels} object or list of objects}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{diskSizeGb}{[Output Only] Size of the snapshot, specified in GB}

\item{id}{[Output Only] The unique identifier for the resource}

\item{labelFingerprint}{A fingerprint for the labels being applied to this snapshot, which is essentially a hash of the labels set used for optimistic locking}

\item{labels}{Labels to apply to this snapshot}

\item{licenses}{[Output Only] A list of public visible licenses that apply to this snapshot}

\item{name}{Name of the resource; provided by the client when the resource is created}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{snapshotEncryptionKey}{Encrypts the snapshot using a customer-supplied encryption key}

\item{sourceDisk}{[Output Only] The source disk used to create this snapshot}

\item{sourceDiskEncryptionKey}{The customer-supplied encryption key of the source disk}

\item{sourceDiskId}{[Output Only] The ID value of the disk used to create this snapshot}

\item{status}{[Output Only] The status of the snapshot}

\item{storageBytes}{[Output Only] A size of the the storage used by the snapshot}

\item{storageBytesStatus}{[Output Only] An indicator whether storageBytes is in a stable state or it is being adjusted as a result of shared storage reallocation}
}
\value{
Snapshot object
}
\description{
Snapshot Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A persistent disk snapshot resource.
}
\seealso{
Other Snapshot functions: \code{\link{Snapshot.labels}},
  \code{\link{disks.createSnapshot}}
}
