% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{SSLHealthCheck}
\alias{SSLHealthCheck}
\title{SSLHealthCheck Object}
\usage{
SSLHealthCheck(port = NULL, portName = NULL, proxyHeader = NULL,
  request = NULL, response = NULL)
}
\arguments{
\item{port}{The TCP port number for the health check request}

\item{portName}{Port name as defined in InstanceGroup#NamedPort#name}

\item{proxyHeader}{Specifies the type of proxy header to append before sending data to the backend, either NONE or PROXY_V1}

\item{request}{The application data to send once the SSL connection has been established (default value is empty)}

\item{response}{The bytes to match against the beginning of the response data}
}
\value{
SSLHealthCheck object
}
\description{
SSLHealthCheck Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
