% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{subnetworks.setIamPolicy}
\alias{subnetworks.setIamPolicy}
\title{Sets the access control policy on the specified resource. Replaces any existing policy.}
\usage{
subnetworks.setIamPolicy(Policy, project, region, resource)
}
\arguments{
\item{Policy}{The \link{Policy} object to pass to this method}

\item{project}{Project ID for this request}

\item{region}{The name of the region for this request}

\item{resource}{Name of the resource for this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute, https://www.googleapis.com/auth/compute.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other Policy functions: \code{\link{Policy}}
}
