% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Tags}
\alias{Tags}
\title{Tags Object}
\usage{
Tags(fingerprint = NULL, items = NULL)
}
\arguments{
\item{fingerprint}{Specifies a fingerprint for this request, which is essentially a hash of the metadata's contents and used for optimistic locking}

\item{items}{An array of tags}
}
\value{
Tags object
}
\description{
Tags Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A set of instance tags.
}
\seealso{
Other Tags functions: \code{\link{instances.setTags}}
}

