% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetHttpProxyList}
\alias{TargetHttpProxyList}
\title{TargetHttpProxyList Object}
\usage{
TargetHttpProxyList(id = NULL, items = NULL, nextPageToken = NULL,
  selfLink = NULL)
}
\arguments{
\item{id}{[Output Only] The unique identifier for the resource}

\item{items}{A list of TargetHttpProxy resources}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] Server-defined URL for this resource}
}
\value{
TargetHttpProxyList object
}
\description{
TargetHttpProxyList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of TargetHttpProxy resources.
}
