% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetHttpsProxiesSetSslCertificatesRequest}
\alias{TargetHttpsProxiesSetSslCertificatesRequest}
\title{TargetHttpsProxiesSetSslCertificatesRequest Object}
\usage{
TargetHttpsProxiesSetSslCertificatesRequest(sslCertificates = NULL)
}
\arguments{
\item{sslCertificates}{New set of SslCertificate resources to associate with this TargetHttpsProxy resource}
}
\value{
TargetHttpsProxiesSetSslCertificatesRequest object
}
\description{
TargetHttpsProxiesSetSslCertificatesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TargetHttpsProxiesSetSslCertificatesRequest functions: \code{\link{targetHttpsProxies.setSslCertificates}}
}

