% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetInstancesScopedList.warning}
\alias{TargetInstancesScopedList.warning}
\title{TargetInstancesScopedList.warning Object}
\usage{

  TargetInstancesScopedList.warning(TargetInstancesScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{TargetInstancesScopedList.warning.data}{The \link{TargetInstancesScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
TargetInstancesScopedList.warning object
}
\description{
TargetInstancesScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Informational warning which replaces the list of addresses when the list is empty.
}
\seealso{
Other TargetInstancesScopedList functions: \code{\link{TargetInstancesScopedList.warning.data}},
  \code{\link{TargetInstancesScopedList}}
}
