% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetPool}
\alias{TargetPool}
\title{TargetPool Object}
\usage{
TargetPool(backupPool = NULL, creationTimestamp = NULL,
  description = NULL, failoverRatio = NULL, healthChecks = NULL,
  id = NULL, instances = NULL, name = NULL, region = NULL,
  selfLink = NULL, sessionAffinity = NULL)
}
\arguments{
\item{backupPool}{This field is applicable only when the containing target pool is serving a forwarding rule as the primary pool, and its failoverRatio field is properly set to a value between [0, 1]}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{failoverRatio}{This field is applicable only when the containing target pool is serving a forwarding rule as the primary pool (i}

\item{healthChecks}{The URL of the HttpHealthCheck resource}

\item{id}{[Output Only] The unique identifier for the resource}

\item{instances}{A list of resource URLs to the virtual machine instances serving this pool}

\item{name}{Name of the resource}

\item{region}{[Output Only] URL of the region where the target pool resides}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{sessionAffinity}{Sesssion affinity option, must be one of the following values:}
}
\value{
TargetPool object
}
\description{
TargetPool Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TargetPool resource. This resource defines a pool of instances, an associated HttpHealthCheck resource, and the fallback target pool.
}
\seealso{
Other TargetPool functions: \code{\link{targetPools.insert}}
}
