% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetPoolAggregatedList.items}
\alias{TargetPoolAggregatedList.items}
\title{TargetPoolAggregatedList.items Object}
\usage{
TargetPoolAggregatedList.items()
}
\value{
TargetPoolAggregatedList.items object
}
\description{
TargetPoolAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] A map of scoped target pool lists.
}
\seealso{
Other TargetPoolAggregatedList functions: \code{\link{TargetPoolAggregatedList}}
}
