% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetPoolInstanceHealth}
\alias{TargetPoolInstanceHealth}
\title{TargetPoolInstanceHealth Object}
\usage{
TargetPoolInstanceHealth(healthStatus = NULL)
}
\arguments{
\item{healthStatus}{No description}
}
\value{
TargetPoolInstanceHealth object
}
\description{
TargetPoolInstanceHealth Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
