% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{targetPools.addHealthCheck}
\alias{targetPools.addHealthCheck}
\title{Adds health check URLs to a target pool.}
\usage{
targetPools.addHealthCheck(TargetPoolsAddHealthCheckRequest, project, region,
  targetPool)
}
\arguments{
\item{TargetPoolsAddHealthCheckRequest}{The \link{TargetPoolsAddHealthCheckRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{region}{Name of the region scoping this request}

\item{targetPool}{Name of the target pool to add a health check to}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other TargetPoolsAddHealthCheckRequest functions: \code{\link{TargetPoolsAddHealthCheckRequest}}
}
