% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{targetPools.removeHealthCheck}
\alias{targetPools.removeHealthCheck}
\title{Removes health check URL from a target pool.}
\usage{
targetPools.removeHealthCheck(TargetPoolsRemoveHealthCheckRequest, project,
  region, targetPool)
}
\arguments{
\item{TargetPoolsRemoveHealthCheckRequest}{The \link{TargetPoolsRemoveHealthCheckRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{region}{Name of the region for this request}

\item{targetPool}{Name of the target pool to remove health checks from}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other TargetPoolsRemoveHealthCheckRequest functions: \code{\link{TargetPoolsRemoveHealthCheckRequest}}
}

