% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{targetPools.setBackup}
\alias{targetPools.setBackup}
\title{Changes a backup target pool's configurations.}
\usage{
targetPools.setBackup(TargetReference, project, region, targetPool,
  failoverRatio = NULL)
}
\arguments{
\item{TargetReference}{The \link{TargetReference} object to pass to this method}

\item{project}{Project ID for this request}

\item{region}{Name of the region scoping this request}

\item{targetPool}{Name of the TargetPool resource to set a backup pool for}

\item{failoverRatio}{New failoverRatio value for the target pool}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other TargetReference functions: \code{\link{TargetReference}},
  \code{\link{forwardingRules.setTarget}},
  \code{\link{globalForwardingRules.setTarget}}
}

