% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetPoolsAddInstanceRequest}
\alias{TargetPoolsAddInstanceRequest}
\title{TargetPoolsAddInstanceRequest Object}
\usage{
TargetPoolsAddInstanceRequest(instances = NULL)
}
\arguments{
\item{instances}{A full or partial URL to an instance to add to this target pool}
}
\value{
TargetPoolsAddInstanceRequest object
}
\description{
TargetPoolsAddInstanceRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TargetPoolsAddInstanceRequest functions: \code{\link{targetPools.addInstance}}
}

