% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetPoolsRemoveInstanceRequest}
\alias{TargetPoolsRemoveInstanceRequest}
\title{TargetPoolsRemoveInstanceRequest Object}
\usage{
TargetPoolsRemoveInstanceRequest(instances = NULL)
}
\arguments{
\item{instances}{URLs of the instances to be removed from target pool}
}
\value{
TargetPoolsRemoveInstanceRequest object
}
\description{
TargetPoolsRemoveInstanceRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TargetPoolsRemoveInstanceRequest functions: \code{\link{targetPools.removeInstance}}
}

