% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetReference}
\alias{TargetReference}
\title{TargetReference Object}
\usage{
TargetReference(target = NULL)
}
\arguments{
\item{target}{No description}
}
\value{
TargetReference object
}
\description{
TargetReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TargetReference functions: \code{\link{forwardingRules.setTarget}},
  \code{\link{globalForwardingRules.setTarget}},
  \code{\link{targetPools.setBackup}}
}

