% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetSslProxiesSetBackendServiceRequest}
\alias{TargetSslProxiesSetBackendServiceRequest}
\title{TargetSslProxiesSetBackendServiceRequest Object}
\usage{
TargetSslProxiesSetBackendServiceRequest(service = NULL)
}
\arguments{
\item{service}{The URL of the new BackendService resource for the targetSslProxy}
}
\value{
TargetSslProxiesSetBackendServiceRequest object
}
\description{
TargetSslProxiesSetBackendServiceRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TargetSslProxiesSetBackendServiceRequest functions: \code{\link{targetSslProxies.setBackendService}}
}
