% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetSslProxiesSetSslCertificatesRequest}
\alias{TargetSslProxiesSetSslCertificatesRequest}
\title{TargetSslProxiesSetSslCertificatesRequest Object}
\usage{
TargetSslProxiesSetSslCertificatesRequest(sslCertificates = NULL)
}
\arguments{
\item{sslCertificates}{New set of URLs to SslCertificate resources to associate with this TargetSslProxy}
}
\value{
TargetSslProxiesSetSslCertificatesRequest object
}
\description{
TargetSslProxiesSetSslCertificatesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TargetSslProxiesSetSslCertificatesRequest functions: \code{\link{targetSslProxies.setSslCertificates}}
}

