% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetVpnGatewayAggregatedList.items}
\alias{TargetVpnGatewayAggregatedList.items}
\title{TargetVpnGatewayAggregatedList.items Object}
\usage{
TargetVpnGatewayAggregatedList.items()
}
\value{
TargetVpnGatewayAggregatedList.items object
}
\description{
TargetVpnGatewayAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of scoped target vpn gateway lists.
}
\seealso{
Other TargetVpnGatewayAggregatedList functions: \code{\link{TargetVpnGatewayAggregatedList}}
}
