% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetVpnGatewayAggregatedList}
\alias{TargetVpnGatewayAggregatedList}
\title{TargetVpnGatewayAggregatedList Object}
\usage{
TargetVpnGatewayAggregatedList(TargetVpnGatewayAggregatedList.items = NULL,
  id = NULL, items = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{TargetVpnGatewayAggregatedList.items}{The \link{TargetVpnGatewayAggregatedList.items} object or list of objects}

\item{id}{[Output Only] The unique identifier for the resource}

\item{items}{A map of scoped target vpn gateway lists}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] Server-defined URL for this resource}
}
\value{
TargetVpnGatewayAggregatedList object
}
\description{
TargetVpnGatewayAggregatedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TargetVpnGatewayAggregatedList functions: \code{\link{TargetVpnGatewayAggregatedList.items}}
}
