% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TestPermissionsRequest}
\alias{TestPermissionsRequest}
\title{TestPermissionsRequest Object}
\usage{
TestPermissionsRequest(permissions = NULL)
}
\arguments{
\item{permissions}{The set of permissions to check for the 'resource'}
}
\value{
TestPermissionsRequest object
}
\description{
TestPermissionsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TestPermissionsRequest functions: \code{\link{addresses.testIamPermissions}},
  \code{\link{autoscalers.testIamPermissions}},
  \code{\link{backendServices.testIamPermissions}},
  \code{\link{disks.testIamPermissions}},
  \code{\link{firewalls.testIamPermissions}},
  \code{\link{forwardingRules.testIamPermissions}},
  \code{\link{globalAddresses.testIamPermissions}},
  \code{\link{globalForwardingRules.testIamPermissions}},
  \code{\link{healthChecks.testIamPermissions}},
  \code{\link{httpHealthChecks.testIamPermissions}},
  \code{\link{httpsHealthChecks.testIamPermissions}},
  \code{\link{images.testIamPermissions}},
  \code{\link{instanceGroupManagers.testIamPermissions}},
  \code{\link{instanceGroups.testIamPermissions}},
  \code{\link{instanceTemplates.testIamPermissions}},
  \code{\link{instances.testIamPermissions}},
  \code{\link{networks.testIamPermissions}},
  \code{\link{regionAutoscalers.testIamPermissions}},
  \code{\link{regionInstanceGroupManagers.testIamPermissions}},
  \code{\link{regionInstanceGroups.testIamPermissions}},
  \code{\link{routers.testIamPermissions}},
  \code{\link{routes.testIamPermissions}},
  \code{\link{snapshots.testIamPermissions}},
  \code{\link{sslCertificates.testIamPermissions}},
  \code{\link{subnetworks.testIamPermissions}},
  \code{\link{targetHttpProxies.testIamPermissions}},
  \code{\link{targetHttpsProxies.testIamPermissions}},
  \code{\link{targetInstances.testIamPermissions}},
  \code{\link{targetPools.testIamPermissions}},
  \code{\link{targetSslProxies.testIamPermissions}},
  \code{\link{targetVpnGateways.testIamPermissions}},
  \code{\link{urlMaps.testIamPermissions}},
  \code{\link{vpnTunnels.testIamPermissions}}
}

