% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TestPermissionsResponse}
\alias{TestPermissionsResponse}
\title{TestPermissionsResponse Object}
\usage{
TestPermissionsResponse(permissions = NULL)
}
\arguments{
\item{permissions}{A subset of `TestPermissionsRequest}
}
\value{
TestPermissionsResponse object
}
\description{
TestPermissionsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

