% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{UDPHealthCheck}
\alias{UDPHealthCheck}
\title{UDPHealthCheck Object}
\usage{
UDPHealthCheck(port = NULL, portName = NULL, request = NULL,
  response = NULL)
}
\arguments{
\item{port}{The UDP port number for the health check request}

\item{portName}{Port name as defined in InstanceGroup#NamedPort#name}

\item{request}{Raw data of request to send in payload of UDP packet}

\item{response}{The bytes to match against the beginning of the response data}
}
\value{
UDPHealthCheck object
}
\description{
UDPHealthCheck Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
