% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{UrlMap}
\alias{UrlMap}
\title{UrlMap Object}
\usage{
UrlMap(creationTimestamp = NULL, defaultService = NULL,
  description = NULL, fingerprint = NULL, hostRules = NULL, id = NULL,
  name = NULL, pathMatchers = NULL, selfLink = NULL, tests = NULL)
}
\arguments{
\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{defaultService}{The URL of the BackendService resource if none of the hostRules match}

\item{description}{An optional description of this resource}

\item{fingerprint}{Fingerprint of this resource}

\item{hostRules}{The list of HostRules to use against the URL}

\item{id}{[Output Only] The unique identifier for the resource}

\item{name}{Name of the resource}

\item{pathMatchers}{The list of named PathMatchers to use against the URL}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{tests}{The list of expected URL mappings}
}
\value{
UrlMap object
}
\description{
UrlMap Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A UrlMap resource. This resource defines the mapping from URL to the BackendService resource, based on the 'longest-match' of the URL's host and path.
}
\seealso{
Other UrlMap functions: \code{\link{urlMaps.insert}},
  \code{\link{urlMaps.patch}}, \code{\link{urlMaps.update}}
}
