% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{UrlMapValidationResult}
\alias{UrlMapValidationResult}
\title{UrlMapValidationResult Object}
\usage{
UrlMapValidationResult(loadErrors = NULL, loadSucceeded = NULL,
  testFailures = NULL, testPassed = NULL)
}
\arguments{
\item{loadErrors}{No description}

\item{loadSucceeded}{Whether the given UrlMap can be successfully loaded}

\item{testFailures}{No description}

\item{testPassed}{If successfully loaded, this field indicates whether the test passed}
}
\value{
UrlMapValidationResult object
}
\description{
UrlMapValidationResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Message representing the validation result for a UrlMap.
}

