% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{vpnTunnels.insert}
\alias{vpnTunnels.insert}
\title{Creates a VpnTunnel resource in the specified project and region using the data included in the request.}
\usage{
vpnTunnels.insert(VpnTunnel, project, region)
}
\arguments{
\item{VpnTunnel}{The \link{VpnTunnel} object to pass to this method}

\item{project}{Project ID for this request}

\item{region}{Name of the region for this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other VpnTunnel functions: \code{\link{VpnTunnel}}
}
