% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{VpnTunnelsScopedList.warning}
\alias{VpnTunnelsScopedList.warning}
\title{VpnTunnelsScopedList.warning Object}
\usage{
VpnTunnelsScopedList.warning(VpnTunnelsScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{VpnTunnelsScopedList.warning.data}{The \link{VpnTunnelsScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
VpnTunnelsScopedList.warning object
}
\description{
VpnTunnelsScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Informational warning which replaces the list of addresses when the list is empty.
}
\seealso{
Other VpnTunnelsScopedList functions: \code{\link{VpnTunnelsScopedList.warning.data}},
  \code{\link{VpnTunnelsScopedList}}
}
